+++ app/Http/Controllers/SolutionController.php
@@ -158,10 +158,14 @@ class SolutionController extends BaseController
     {
         $solution = $this->solution_repo->find($solution_id);
 
+        $image_gallery = json_decode($solution->image_gallery);
+        $image_gallery = str_replace('\n', '', json_encode($image_gallery, JSON_HEX_QUOT | JSON_HEX_TAG | JSON_HEX_AMP | JSON_HEX_APOS | JSON_PRESERVE_ZERO_FRACTION));
+
         return view('solution.detail')->with([
             'is_restricted'    => $this->is_restricted_adminer,
             'project_tag_tree' => $this->project_repo->projectTagTree(),
             'solution'         => $solution,
+            'image_gallery'    => $image_gallery
         ]);
     }
 
@@ -176,6 +180,9 @@ class SolutionController extends BaseController
         $solution = $is_wait_approve_ongoing ?
             $this->solution_repo->findDuplicate($solution_id) : $this->solution_repo->find($solution_id);
 
+        $image_gallery = json_decode($solution->image_gallery);
+        $image_gallery = str_replace('\n', '', json_encode($image_gallery, JSON_HEX_QUOT | JSON_HEX_TAG | JSON_HEX_AMP | JSON_HEX_APOS | JSON_PRESERVE_ZERO_FRACTION));
+
         return view($is_wait_approve_ongoing ? 'solution.update-ongoing' : 'solution.update')
             ->with([
                 'is_restricted'            => $this->is_restricted_adminer,
@@ -188,6 +195,7 @@ class SolutionController extends BaseController
                 'project_category_options' => $this->project_repo->categoryOptions(),
 
                 'solution'                 => $solution,
+                'image_gallery'            => $image_gallery,
             ]);
     }