+++ resources/views/solution/detail.blade.php
@@ -36,7 +36,7 @@
 			<div id="solution-gallery" class="solution-thumbs"
 				 data-mode="display"
 				 data-solution-cover="{!! $solution->image !!}"
-				 data-solution-gallery='{!! $solution->image_gallery !!}'>
+				 data-solution-gallery='{!! $image_gallery !!}'>
 			</div>
         </div>
         @endforeach