+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ## v1.5.8
+* [FIXED] fix JSON.parse parse quote error from solution edit image (@Hank)
 * [FIXED] fix solution edit, array not found thumb_delete_x index key error (@Hank)
 
 ## v1.5.7